refactored int to double

+++ src/main/java/instruments/Guitar.java
@@ -5,7 +5,7 @@ public class Guitar extends Instrument{
     private int numOfStrings;
 
 
-    public Guitar(String type, String colour, int stockPrice, int sellPrice, int numOfStrings){
+    public Guitar(String type, String colour, double stockPrice, double sellPrice, int numOfStrings){
         super(type, colour, stockPrice, sellPrice);
         this.numOfStrings = numOfStrings;
     }
